requirements.txt
numpy
quickstart-vdk
vdk-trino


# necessary for sql in jupyter
ipython-sql
jupyterlab_sql

streamlit
jupyter-server-proxy
nbserverproxy

pyarrow==7.0.0
seaborn
matplotlib
scikit-learn
xlsxwriter